refactor: reorganize code.
